tests/wpt/web-platform-tests/old-tests/submission/Microsoft/sandbox/sandbox_002.htm

<!DOCTYPE html>
<html>
<head>
    <title>HTML5 Sandbox: Allow autoplay for HTML5 Video inside iframe with sandbox attribute if sandbox='allow-scripts'.</title>
    <meta name=timeout content=long>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type" />
    <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script>
      async_test(function (t) {
        var callback = t.step_func_done(function(event) {
          assert_true('sandbox' in document.createElement('iframe'));
          assert_equals(event.data, "play event fired");
        });

        window.addEventListener("message", callback, false);
      }, "Allow autoplay for HTML5 Video inside iframe with sandbox attribute if sandbox='allow-scripts'.");
    </script>
    <iframe src="iframe_sandbox_002.htm" sandbox="allow-scripts" style="display: none"></iframe>
    <div id=log></div>
</body>
</html>
